Elevate GitHubResponse.ResponseInfo into its own public class.
